rag/rag_redis_adk_pdf_upload/requirements.txt
# ── Core framework ──
google-adk
streamlit

# ── Redis vector store ──
redis
redisvl

# ── PDF processing ──
pdfplumber
pypdf

# ── Embeddings ──
sentence-transformers

# ── Utilities ──
python-dotenv
numpy
torchvision